requirements.txt
streamlit>=1.35.0
llama-index>=0.10.0
llama-index-llms-groq>=0.1.4
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-chroma>=0.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pillow>=10.3.0
easyocr>=1.7.1
opencv-python-headless>=4.9.0
moviepy>=1.0.3
SpeechRecognition>=3.10.0
pydub>=0.25.1
youtube-transcript-api>=0.6.1
yt-dlp>=2024.4.9
chromadb>=0.5.0
sentence-transformers>=3.0.0
groq>=0.9.0
tiktoken>=0.7.0
numpy>=1.26.0
setuptools>=65.5.0
